requirements.txt
torch
torchvision
transformers
scikit-learn
opencv-python
flask
python-dotenv
google-genai
